internet-testing/requirements.txt
streamlit
pandas
matplotlib
numpy
python-nmap
PyYAML
kubernetes